Moves table code into mixins
